pythonSharePrice/app/requirements.txt
Flask
mysql-connector
Flask-HTTPAuth
Flask-BasicAuth
